not to interfaces...
