+++ components/Journey.vue
@@ -3,8 +3,8 @@
   <div class="sequence">
     <div class="scene" :class="activeSceneClasses">
       <div class="main-visual-container" :style="mainVisualContainerStyles">
-        <main-visual :canvas="canvas" :mainVisual="mainVisual"></main-visual>
-        <visual-tags :sequence="sequence" :activeScene="activeScene" :canvas.sync="canvas" :actual="actual" :mainVisual="mainVisual"></visual-tags>
+        <main-visual :canvasIsLarger="canvasIsLarger" :mainVisual="mainVisual"></main-visual>
+        <visual-tags :canvas.sync="canvas" :tags="activeScene.visualTags" :actual="actual" :mainVisual="mainVisual" :getStyles="getStyles"></visual-tags>
         <div class="subtitle d-flex justify-content-center">
           <subtitling-machine :lines="activeScene.subtitle" :config="subtitling.config" :status.sync="subtitling.status" :key="activeScene.id" />
         </div>
@@ -42,11 +42,10 @@ import { knowsMarkdown, knowsDOM, knowsAudio } from 'watchout-common-functions/i
 import MainVisual from './journey/another-future/MainVisual'
 import VisualTags from './journey/another-future/VisualTags'
 import SubtitlingMachine from './journey/SubtitlingMachine'
-import getStyles from '~/interfaces/journey/getStyles'
-import canvasIsLarger from '~/interfaces/journey/canvasIsLarger'
+import parseColor from 'parse-color'
 
 export default {
-  mixins: [knowsMarkdown, knowsDOM, knowsAudio, getStyles, canvasIsLarger],
+  mixins: [knowsMarkdown, knowsDOM, knowsAudio],
   props: ['project'],
   data() {
     var state = {
@@ -153,6 +152,50 @@ export default {
     }
   },
   methods: {
+    canvasIsLarger() {
+      return this.canvas.width >= this.mainVisual.width && this.canvas.height >= this.mainVisual.height
+    },
+    getStyles(name, data = undefined) {
+      var styles = {}
+      const global = this.sequence.default ? this.sequence.default.styles ? this.sequence.default.styles[name] : undefined : undefined
+      const scene = this.activeScene.default ? this.activeScene.default.styles ? this.activeScene.default.styles[name] : undefined : undefined
+      const local = data ? data.styles : undefined
+      const attributes = ['text', 'border', 'background']
+      for(let attribute of attributes) {
+        styles[attribute] = (local && local[attribute]) || (scene && scene[attribute]) || (global && global[attribute]) || undefined
+      }
+
+      if(styles.text && typeof styles.text === 'object') {
+        if(styles.text.align) {
+          styles.textAlign = styles.text.align
+        }
+        if(styles.text.size) {
+          styles.fontSize = styles.text.size
+        }
+        if(styles.text.color) {
+          styles.color = styles.text.color
+        }
+        if(styles.text.leading) {
+          styles.lineHeight = styles.text.leading
+        }
+      }
+      if(styles.border && typeof styles.border === 'object') {
+        if(styles.border.width) {
+          styles.borderWidth = styles.border.width
+        }
+        if(styles.border.color) {
+          styles.borderColor = styles.border.color
+        }
+      }
+      if(styles.background && typeof styles.background === 'object') {
+        var color = parseColor(styles.background.color)
+        if(styles.background.opacity) {
+          color = parseColor('rgba(' + color.rgb.slice().concat(styles.background.opacity).join(',') + ')')
+        }
+        styles.backgroundColor = 'rgba(' + color.rgba.join(',') + ')'
+      }
+      return styles
+    },
     getMediaFromID(id) {
       return Object.assign({}, this.sequence.media.find(media => media.id === id))
     },
