--- interfaces/journey/canvasIsLarger.vue
@@ -1,9 +1,0 @@
-<script>
-export default {
-  methods: {
-    canvasIsLarger() {
-      return this.canvas.width >= this.mainVisual.width && this.canvas.height >= this.mainVisual.height
-    }
-  }
-}
-</script>
